Added a reference to the unit system inside the Unit class
This means that ut_free_system is only called by the Python GC *after* all
Units have been deallocated with ut_free. Before this change, it was
possible for ut_free_system to have been called first, and the result was that
subsequent ut_free calls resulted in a Segmentation Fault 11.